Rewrite a lot,fix memory lake and add an dialog.

+++ app/src/main/java/net/fengberd/minecraftpe_server/ServerUtils.java
@@ -1,33 +1,34 @@
 package net.fengberd.minecraftpe_server;
 
+import android.content.*;
+import android.os.*;
+
 import java.io.*;
-import java.util.*;
-import java.nio.charset.Charset;
+import java.lang.Process;
 
-import android.content.Context;
-
-public final class ServerUtils
+public class ServerUtils
 {
-	public static Context mContext;
+	private static File appDirectory=null;
+	private static File nukkitDataDirectory=new File(Environment.getExternalStorageDirectory(),"Nukkit"),pocketmineDataDirectory=new File(Environment.getExternalStorageDirectory(),"PocketMine");
 
-	private static Process serverProcess;
-	private static OutputStream stdin;
-	private static InputStream stdout;
+	private static Process serverProcess=null;
+	private static InputStreamReader stdout=null;
+	private static OutputStreamWriter stdin=null;
 
-	public static void setContext(Context mContext)
+	public static void setAppDirectory(Context ctx)
 	{
-		ServerUtils.mContext=mContext;
+		appDirectory=ctx.getFilesDir().getParentFile();
 	}
 
-	public static String getAppDirectory()
+	public static File getAppDirectory()
 	{
-		return mContext.getApplicationInfo().dataDir;
+		return appDirectory;
 	}
 
-	public static String getDataDirectory()
+	public static File getDataDirectory()
 	{
-		String dir=android.os.Environment.getExternalStorageDirectory().getPath() + (MainActivity.nukkitMode?"/Nukkit":"/PocketMine");
-		new File(dir).mkdirs();
+		File dir=MainActivity.nukkitMode ? nukkitDataDirectory : pocketmineDataDirectory;
+		dir.mkdirs();
 		return dir;
 	}
 
@@ -35,15 +36,18 @@ public final class ServerUtils
 	{
 		try
 		{
-			Runtime.getRuntime().exec(getAppDirectory() + "/busybox killall -9 " +(MainActivity.nukkitMode?"java":"php")).waitFor();
+			// TODO: This might kill other processes?
+			Runtime.getRuntime()
+					.exec(getAppDirectory() + "/busybox killall -9 " + (MainActivity.nukkitMode ? "java" : "php"))
+					.waitFor();
 		}
-		catch(Exception e)
+		catch(Exception ignored)
 		{
-			
+
 		}
 	}
 
-	public static Boolean isRunning()
+	public static boolean isRunning()
 	{
 		try
 		{
@@ -56,31 +60,17 @@ public final class ServerUtils
 		return false;
 	}
 
-	final public static void runServer()
+	public static void runServer()
 	{
-		File f = new File(getDataDirectory(),"/tmp");
-		if(!f.exists())
-		{
-			f.mkdir();
-		}
-		else if(!f.isDirectory())
+		File f=new File(getDataDirectory(),"tmp");
+		if(!f.isDirectory() && f.exists())
 		{
 			f.delete();
-			f.mkdir();
 		}
+		f.mkdirs();
 		setPermission();
-		String file = "/Nukkit.jar";
-		if(!MainActivity.nukkitMode)
-		{
-			if(new File(getDataDirectory() + "/PocketMine-MP.phar").exists())
-			{
-				file="/PocketMine-MP.phar";
-			}
-			else
-			{
-				file = "/src/pocketmine/PocketMine.php";
-			}
-		}
+		String file=MainActivity.nukkitMode ? "/Nukkit.jar" : (new File(getDataDirectory(),"/PocketMine-MP.phar")
+				.exists() ? "/PocketMine-MP.phar" : "/src/pocketmine/PocketMine.php");
 		File ini=new File(getDataDirectory() + "/php.ini");
 		if(!MainActivity.nukkitMode && !ini.exists())
 		{
@@ -88,10 +78,11 @@ public final class ServerUtils
 			{
 				ini.createNewFile();
 				FileOutputStream os=new FileOutputStream(ini);
-				os.write("phar.readonly=0\nphar.require_hash=1\ndate.timezone=Asia/Shanghai\nshort_open_tag=0\nasp_tags=0\nopcache.enable=1\nopcache.enable_cli=1\nopcache.save_comments=1\nopcache.fast_shutdown=0\nopcache.max_accelerated_files=4096\nopcache.interned_strings_buffer=8\nopcache.memory_consumption=128\nopcache.optimization_level=0xffffffff".getBytes("UTF8"));
+				os.write("phar.readonly=0\nphar.require_hash=1\ndate.timezone=Asia/Shanghai\nshort_open_tag=0\nasp_tags=0\nopcache.enable=1\nopcache.enable_cli=1\nopcache.save_comments=1\nopcache.fast_shutdown=0\nopcache.max_accelerated_files=4096\nopcache.interned_strings_buffer=8\nopcache.memory_consumption=128\nopcache.optimization_level=0xffffffff"
+						.getBytes("UTF8"));
 				os.close();
 			}
-			catch(Exception e)
+			catch(Exception ignored)
 			{
 
 			}
@@ -99,77 +90,74 @@ public final class ServerUtils
 		String[] args=null;
 		if(MainActivity.nukkitMode)
 		{
-			args=new String[]
-			{
-				getAppDirectory() + "/java/jre/bin/java",
-				"-jar",
-				getDataDirectory() + file,
-				MainActivity.ansiMode?"enable-ansi":"disable-ansi",
-				"disable-jline"
+			args=new String[]{
+					getAppDirectory() + "/java/jre/bin/java",
+					"-jar",
+					getDataDirectory() + file,
+					MainActivity.ansiMode ? "enable-ansi" : "disable-ansi",
+					"disable-jline"
 			};
 		}
 		else
 		{
-			args=new String[]
-			{
-				getAppDirectory() + "/php",
-				"-c",
-				getDataDirectory() + "/php.ini",
-				getDataDirectory() + file,
-				MainActivity.ansiMode?"--enable-ansi":"--disable-ansi"
+			args=new String[]{
+					getAppDirectory() + "/php",
+					"-c",
+					getDataDirectory() + "/php.ini",
+					getDataDirectory() + file,
+					MainActivity.ansiMode ? "--enable-ansi" : "--disable-ansi"
 			};
 		}
-		ProcessBuilder builder = new ProcessBuilder(args);
+		ProcessBuilder builder=new ProcessBuilder(args);
 		builder.redirectErrorStream(true);
-		builder.directory(new File(getDataDirectory()));
-		Map<String,String> env=builder.environment();
-		env.put("TMPDIR",getDataDirectory() + "/tmp");
+		builder.directory(getDataDirectory());
+		builder.environment().put("TMPDIR",getDataDirectory() + "/tmp");
 		try
 		{
 			serverProcess=builder.start();
-			stdout=serverProcess.getInputStream();
-			stdin=serverProcess.getOutputStream();
-			Thread tMonitor = new Thread()
+			stdout=new InputStreamReader(serverProcess.getInputStream(),"UTF-8");
+			stdin=new OutputStreamWriter(serverProcess.getOutputStream(),"UTF-8");
+			Thread tMonitor=new Thread()
 			{
 				public void run()
 				{
-					InputStreamReader reader = new InputStreamReader(stdout,Charset.forName("UTF-8"));
-					BufferedReader br = new BufferedReader(reader);
+					BufferedReader br=new BufferedReader(stdout);
 					while(isRunning())
 					{
 						try
 						{
-							char[] buffer = new char[8192];
-							int size = 0;
-							while((size = br.read(buffer,0,buffer.length)) != -1)
+							int size=0;
+							char[] buffer=new char[8192];
+							StringBuilder s=new StringBuilder();
+							while((size=br.read(buffer,0,buffer.length))!=-1)
 							{
-								StringBuilder s = new StringBuilder();
-								for(int i = 0; i < size; i++) 
+								s.setLength(0);
+								for(int i=0;i<size;i++)
 								{
-									char c = buffer[i];
-									if(c == '\r')
+									char c=buffer[i];
+									switch(c)
 									{
+									case '\r':
 										continue;
-									}
-									if(c == '\n' || c == '\u0007')
-									{
-										String line = s.toString();
-										if(c == '\u0007' || line.startsWith("\u001B]0;"))
-										{
-											//Do nothing.
-										}
-										else
+									case '\n':
+										String line=s.toString();
+										if(!line.startsWith("\u001B]0;"))
 										{
 											ConsoleActivity.log(line);
 										}
-										s=new StringBuilder();
-									}
-									else
-									{
-										s.append(buffer[i]);
+									case '\u0007':
+										s.setLength(0);
+										break;
+									default:
+										s.append(c);
+										break;
 									}
 								}
 							}
+						}
+						catch(IOException ignored)
+						{
+
 						}
 						catch(Exception e)
 						{
@@ -188,51 +176,52 @@ public final class ServerUtils
 						}
 					}
 					ConsoleActivity.log("[PE Server] Server was stopped.");
-					MainActivity.stopNotifyService();
+					MainActivity.postMessage(MainActivity.ACTION_STOP_SERVICE,0,null);
 				}
 			};
 			tMonitor.start();
 		}
 		catch(Exception e)
 		{
-			ConsoleActivity.log("[PE Server] Unable to start "+(MainActivity.nukkitMode?"Java":"PHP")+".");
+			ConsoleActivity.log("[PE Server] Unable to start " + (MainActivity.nukkitMode ? "Java" : "PHP") + ".");
 			ConsoleActivity.log(e.toString());
-			MainActivity.stopNotifyService();
+			MainActivity.postMessage(MainActivity.ACTION_STOP_SERVICE,0,null);
 			killServer();
 		}
-		return;
 	}
 
-	final static public void setPermission()
+	public static void setPermission()
 	{
 		try
 		{
 			if(MainActivity.nukkitMode)
 			{
-				Runtime.getRuntime().exec("./busybox chmod -R 755 java",new String[0],new File(getAppDirectory())).waitFor();
+				Runtime.getRuntime()
+						.exec("./busybox chmod -R 755 java",new String[0],appDirectory)
+						.waitFor();
 			}
 			else
 			{
-				new File(getAppDirectory()+"/php").setExecutable(true,true);
+				new File(appDirectory,"php").setExecutable(true,true);
 			}
 		}
 		catch(Exception e)
 		{
-
+			e.printStackTrace();
 		}
 	}
 
-	public static void writeCommand(String Cmd)
+	public static boolean writeCommand(String cmd)
 	{
 		try
 		{
-			stdin.write((Cmd + "\r\n").getBytes());
+			stdin.write(cmd + "\r\n");
 			stdin.flush();
 		}
 		catch(Exception e)
 		{
-
+			return false;
 		}
+		return true;
 	}
 }
-
